(Simulation)Vehicles travel at constant speed on a flat road.For this reason, minor changes were made in the classes.

+++ src/main/scala/Model/Velocity.scala
@@ -1,14 +1,14 @@
 package Model
 
-case class Acceleration(ax: Double, ay: Double){
+case class Acceleration(ax: Float, ay: Float){
   var accelerationTotal = Math.sqrt(ax*ax + ay*ay)
 }
 
-case class Velocity(var dx: Double, var dy: Double) {
+case class Velocity(var dx: Float, var dy: Float) {
   require(dx >= 0)
   require(dy >= 0)
 
-  lazy val vitesseTotal: Double = Math.sqrt(dx * dx + dy * dy)
+  lazy val vitesseTotal: Float = Math.sqrt(dx * dx + dy * dy).toFloat
 
   def moveFrom(from: Position): Position =
     from.move(dx, dy)
